requirements.txt
pandas
numpy
requests
bs4
mysql-connector-python